Initialized for WROOM board
If you look at the article the board they reference is "ESP32 DevKit".
I'm no ESP32 conniseaur (I wonder if I can get a spellcheck in here) but
I do think that has some different configs than other ESP32's.

That said, the board in the pictures is definitely the *ESP-WROOM-32*.
I HAVE NO IDEA if this will be significantly different for us or not.
You may even be interested in a different board.

For now until we have hardware this will be the branch this lives on
with the assumption that *ESP-WROOM-32* is what we're rocking with.

I set this up with PlatformIO. This I guess is like XCode for iOS but
for embedded systems. Not sure! But it helps us stay in whatever editors
we like.

Ref:
[PlatformIO for VSCode](https://www.youtube.com/watch?v=PYSy_PLjytQ)