Testing with Uno
Since I have my Arduino Uno it was easiest to test this platformIO stuff
using this without mixing with whatever env we'd need to set up for the
other board we get.

Most concerning is that my text editor (nvim btw) does *not* fw this
arduino stuff. Not a lot of people using both! Got it to stop bitching
with a compiled_commands.json [reading
this](https://community.platformio.org/t/using-vscode-clangd-fails-to-locate-arduino-h/3384/8). This probably isn't very group project friendly unforch...

Will wanna figure something else out but this works for now